+++ packages/grz-common/src/grz_common/__init__.py
@@ -2,4 +2,4 @@
 Common library for grz applications.
 """
 
-__version__ = "1.5.0"
+__version__ = "1.6.0"